BAP V1.0.28 - finish process WorkFusion
Since I'm spending way too much time on this and getting nowhere, it's labeld as incompletable. This means that the script is written but it doesn't work and I can't get it working. If there is any time left, this will get picked up again but for now, it's finished